move estimateGas to internals
